Fix: Friction

+++ Game/Server/Physics.h
@@ -22,7 +22,7 @@ public:
 	~Physics();
 
 
-	void CaculateVelocity(PlayerInfo* players, double elapsedTime);
+	void CaculateVelocity(PlayerInfo* players, float elapsedTime);
 	void CaculatePosition(PlayerInfo* players);
 
 	void AfterColideWithWall(float* Acc, float* Vel, float diff);
